requirements.txt
google-genai==0.3.0
python-dotenv==1.0.1
Pillow==10.4.0
fastapi==0.112.2
uvicorn==0.30.6
elevenlabs==1.7.0
fal==2.5.0
httpx==0.27.2
requests==2.32.3

